medusa/engine/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "medusa-engine"
version = "1.0.0"
description = "Medusa security research framework engine"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.22.0",
    "httpx>=0.24.0",
    "beautifulsoup4>=4.12.0",
    "pydantic>=2.0.0",
    "sqlalchemy>=2.0.0",
    "pyyaml>=6.0",
    "litellm>=1.40.0",
    "python-json-logger>=2.0.0",
    "websockets>=12.0",
]

[tool.setuptools.packages.find]
where = [".."]
include = ["medusa*"]
